requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.1
pydantic>=2.5.3
pydantic-settings>=2.1.0
requests>=2.31.0
pdfplumber>=0.10.3
python-docx>=1.1.0
chromadb>=0.4.22
sentence-transformers>=2.3.1
openai>=1.0.0
streamlit>=1.28.0


